Split merged utils.py into per-task independent utils
Replace the single 2900-line merged utils.py with the original
per-task utils.py files from the reference implementation. Each
sub-task folder now has its own independent utils.py, eliminating
function name collisions entirely.
