+++ lmms_eval/tasks/unig2u/babyvision/fine_grained_discrimination.yaml
@@ -8,11 +8,11 @@ output_type: generate_until
 
 process_docs: !function utils.process_fine_grained
 
-doc_to_visual: !function utils.babyvision_doc_to_visual
-doc_to_text: !function utils.babyvision_doc_to_text
+doc_to_visual: !function utils.doc_to_visual
+doc_to_text: !function utils.doc_to_text
 doc_to_target: "blankAns"
 
-process_results: !function utils.babyvision_process_results
+process_results: !function utils.process_results
 
 lmms_eval_specific_kwargs:
   default:
@@ -26,7 +26,7 @@ generation_kwargs:
 
 metric_list:
   - metric: Fine-grained Discrimination
-    aggregation: !function utils.babyvision_aggregate_results
+    aggregation: !function utils.aggregate_results
     higher_is_better: true
 
 metadata:
