--- lmms_eval/tasks/unig2u/illusionbench/utils.py
@@ -1,1 +1,0 @@
-../utils.py
