+++ lmms_eval/tasks/unig2u/mmsi/utils.py
@@ -1,1 +1,148 @@
-../utils.py
+import io
+import logging
+import re
+from collections import defaultdict
+
+import numpy as np
+import pandas as pd
+from PIL import Image
+
+from lmms_eval.filters.extraction import ExtendedRegexFilter
+from lmms_eval.filters.transformation import MapFilter
+
+eval_logger = logging.getLogger("lmms-eval")
+
+
+def msr_doc_to_text(doc, lmms_eval_specific_kwargs=None):
+    question = doc["question"].strip()
+    if "pre_prompt" in lmms_eval_specific_kwargs and lmms_eval_specific_kwargs["pre_prompt"] != "":
+        question = f"{lmms_eval_specific_kwargs['pre_prompt']}{question}"
+    if "post_prompt" in lmms_eval_specific_kwargs and lmms_eval_specific_kwargs["post_prompt"] != "":
+        question = f"{question}{lmms_eval_specific_kwargs['post_prompt']}"
+    return question
+
+
+def msr_doc_to_text_with_gen_prompt(doc, lmms_eval_specific_kwargs=None):
+    """
+    Version for visual CoT models that includes generation prompt in a parseable format.
+
+    Output format: [GEN_PROMPT]...[/GEN_PROMPT][QUESTION]...[/QUESTION]
+    - Stage 1 uses content inside [GEN_PROMPT] tags
+    - Stage 2 uses content inside [QUESTION] tags (includes pre_prompt and post_prompt)
+    """
+    question = doc["question"].strip()
+
+    # Build Stage 2 question with pre_prompt and post_prompt
+    stage2_question = question
+    if "pre_prompt" in lmms_eval_specific_kwargs and lmms_eval_specific_kwargs["pre_prompt"] != "":
+        stage2_question = f"{lmms_eval_specific_kwargs['pre_prompt']}{stage2_question}"
+    if "post_prompt" in lmms_eval_specific_kwargs and lmms_eval_specific_kwargs["post_prompt"] != "":
+        stage2_question = f"{stage2_question}{lmms_eval_specific_kwargs['post_prompt']}"
+
+    # Add generation prompt as a special marker if provided
+    if "generation_prompt" in lmms_eval_specific_kwargs:
+        gen_prompt = lmms_eval_specific_kwargs["generation_prompt"]
+        # Format: [GEN_PROMPT]...[/GEN_PROMPT][QUESTION]...[/QUESTION]
+        return f"[GEN_PROMPT]{gen_prompt}[/GEN_PROMPT][QUESTION]{stage2_question}[/QUESTION]"
+
+    return stage2_question
+
+
+def msr_doc_to_visual(doc):
+    image_list = []
+    for img_data in doc["images"]:
+        # Check if already a PIL Image object
+        if isinstance(img_data, Image.Image):
+            image = img_data.convert("RGB")
+        elif isinstance(img_data, dict):
+            # If dict (from HuggingFace datasets), extract bytes
+            image = Image.open(io.BytesIO(img_data["bytes"]))
+            image = image.convert("RGB")
+        else:
+            # If bytes data, decode it
+            image = Image.open(io.BytesIO(img_data))
+            image = image.convert("RGB")
+        image_list.append(image)
+    return image_list
+
+
+def extract_single_choice_with_word_boundary(pred, gt):
+    pattern_1 = r"``([^`]*)``"
+    match = re.search(pattern_1, pred)
+    if match:
+        pred = match.group(1)
+
+    pattern_2 = r"`([^`]*)`"
+    match = re.search(pattern_2, pred)
+    if match:
+        pred = match.group(1)
+
+    pattern_add = r"\{([^}]*)\}"
+    match = re.search(pattern_add, pred)
+    if match:
+        pred = match.group(1)
+
+    pattern_3 = r"\b[A-D]\b(?!\s[a-zA-Z])"
+    match = re.search(pattern_3, pred)
+    if match:
+        pred = match.group()
+    else:
+        return None
+
+    answer = gt.lower().replace("\n", " ").strip()
+    predict = pred.lower().replace("\n", " ").strip()
+    try:
+        if answer == predict[0]:
+            return 1.0
+        elif predict[0] == "(" and answer == predict[1]:
+            return 1.0
+        elif predict[0:7] == "option " and answer == predict[7]:
+            return 1.0
+        elif predict[0:14] == "the answer is " and answer == predict[14]:
+            return 1.0
+    except Exception as e:
+        return 0.0
+    return 0.0
+
+
+def msr_process_results(doc, results):
+    """
+    Args:
+        doc: a instance of the eval dataset
+        results: [pred]
+    Returns:
+        a dictionary with key: metric name, value: metric value
+    """
+    pred = results[0]
+    gt = doc["answer"]
+
+    score = extract_single_choice_with_word_boundary(pred, gt)
+    category = doc["question_type"]
+    l2_category = doc["question_type"]
+    if score is None:
+        return {category: {"question_id": doc["id"], "l2_category": l2_category, "score": 0, "note": "can not find anwser"}, "average": {"question_id": doc["id"], "l2_category": l2_category, "score": 0, "note": "can not find anwser"}}
+    return {category: {"question_id": doc["id"], "l2_category": l2_category, "score": score}, "average": {"question_id": doc["id"], "l2_category": l2_category, "score": score}}
+
+
+def msr_aggregate_results(results):
+    """
+    Args:
+        results: a list of values returned by process_results
+    Returns:
+        A score
+    """
+    l2_category_scores = defaultdict(list)
+    for result in results:
+        score = result["score"]
+        l2_category = result["l2_category"]
+        l2_category_scores[l2_category].append(score)
+
+    l2_category_avg_score = {}
+    for l2_category, scores in l2_category_scores.items():
+        avg_score = sum(scores) / len(scores)
+        l2_category_avg_score[l2_category] = avg_score
+        eval_logger.info(f"{l2_category}: {avg_score:.2f}")
+
+    all_scores = [score for scores in l2_category_scores.values() for score in scores]
+    avg_score = sum(all_scores) / len(all_scores) if all_scores else 0.0
+    return avg_score
