+++ lmms_eval/tasks/unig2u/realunify/attentional_focusing.yaml
@@ -6,11 +6,11 @@ task: "unig2u_realunify_attentional_focusing"
 test_split: train
 output_type: generate_until
 
-doc_to_visual: !function utils.realunify_doc_to_visual
-doc_to_text: !function utils.realunify_doc_to_text
+doc_to_visual: !function utils.doc_to_visual
+doc_to_text: !function utils.doc_to_text
 doc_to_target: "answer"
 
-process_results: !function utils.realunify_process_results
+process_results: !function utils.process_results
 
 lmms_eval_specific_kwargs:
   default:
@@ -24,7 +24,7 @@ generation_kwargs:
 
 metric_list:
   - metric: attentional_focusing
-    aggregation: !function utils.realunify_aggregate_results
+    aggregation: !function utils.aggregate_results
     higher_is_better: true
 
 metadata:
